use dumb header component
