+++ packages/core/admin/admin/src/pages/MarketplacePage/index.js
@@ -13,23 +13,23 @@ import {
   useAppInfos,
 } from '@strapi/helper-plugin';
 import { Grid, GridItem } from '@strapi/design-system/Grid';
-import { Layout, HeaderLayout, ContentLayout } from '@strapi/design-system/Layout';
+import { Layout, ContentLayout } from '@strapi/design-system/Layout';
 import { Main } from '@strapi/design-system/Main';
 import { Searchbar } from '@strapi/design-system/Searchbar';
 import { Box } from '@strapi/design-system/Box';
-import { LinkButton } from '@strapi/design-system/LinkButton';
 import { useNotifyAT } from '@strapi/design-system/LiveRegions';
 import { Typography } from '@strapi/design-system/Typography';
 import { Flex } from '@strapi/design-system/Flex';
-import Upload from '@strapi/icons/Upload';
 
 import PluginCard from './components/PluginCard';
 import { EmptyPluginSearch } from './components/EmptyPluginSearch';
+import MarketplacePageHeader from './components/MarketplacePageHeader';
 import { fetchAppInformation } from './utils/api';
 import useFetchInstalledPlugins from '../../hooks/useFetchInstalledPlugins';
 import useFetchMarketplacePlugins from '../../hooks/useFetchMarketplacePlugins';
 import adminPermissions from '../../permissions';
 import offlineCloud from '../../assets/images/icon_offline-cloud.svg';
+import useNavigatorOnLine from './hooks/use-navigator-on-line';
 
 const matchSearch = (plugins, search) => {
   return matchSorter(plugins, search, {
@@ -41,30 +41,6 @@ const matchSearch = (plugins, search) => {
       { threshold: matchSorter.rankings.WORD_STARTS_WITH, key: 'attributes.description' },
     ],
   });
-};
-
-const getOnLineStatus = () =>
-  typeof navigator !== 'undefined' && typeof navigator.onLine === 'boolean'
-    ? navigator.onLine
-    : true;
-
-const useNavigatorOnLine = () => {
-  const [status, setStatus] = React.useState(getOnLineStatus());
-
-  const setOnline = () => setStatus(true);
-  const setOffline = () => setStatus(false);
-
-  useEffect(() => {
-    window.addEventListener('online', setOnline);
-    window.addEventListener('offline', setOffline);
-
-    return () => {
-      window.removeEventListener('online', setOnline);
-      window.removeEventListener('offline', setOffline);
-    };
-  }, []);
-
-  return status;
 };
 
 const MarketPlacePage = () => {
@@ -147,30 +123,7 @@ const MarketPlacePage = () => {
   if (!isOnline) {
     return (
       <Layout>
-        <HeaderLayout
-          title={formatMessage({
-            id: 'global.marketplace',
-            defaultMessage: 'Marketplace',
-          })}
-          subtitle={formatMessage({
-            id: 'admin.pages.MarketPlacePage.subtitle',
-            defaultMessage: 'Get more out of Strapi',
-          })}
-          primaryAction={
-            <LinkButton
-              startIcon={<Upload />}
-              variant="tertiary"
-              href="https://market.strapi.io/submit-plugin"
-              onClick={() => trackUsage('didSubmitPlugin')}
-            >
-              {formatMessage({
-                id: 'admin.pages.MarketPlacePage.submit.plugin.link',
-                defaultMessage: 'Submit your plugin',
-              })}
-            </LinkButton>
-          }
-        />
-
+        <MarketplacePageHeader />
         <Flex
           width="100%"
           direction="column"
@@ -178,14 +131,20 @@ const MarketPlacePage = () => {
           justifyContent="center"
           paddingTop={11}
         >
-          <Box paddingTop={8} paddingBottom={1}>
+          <Box paddingTop={11} paddingBottom={1}>
             <Typography textColor="neutral700" variant="alpha">
-              You are offline
+              {formatMessage({
+                id: 'admin.pages.MarketPlacePage.offline.title',
+                defaultMessage: 'You are offline',
+              })}
             </Typography>
           </Box>
           <Box paddingBottom={3}>
             <Typography textColor="neutral700" variant="epsilon">
-              You need to be connected to the Internet to access Strapi Market.
+              {formatMessage({
+                id: 'admin.pages.MarketPlacePage.offline.subtitle',
+                defaultMessage: 'You need to be connected to the Internet to access Strapi Market.',
+              })}
             </Typography>
           </Box>
           <Box as="img" src={offlineCloud} alt="offline" width={11} height={11} />
@@ -228,29 +187,7 @@ const MarketPlacePage = () => {
             defaultMessage: 'Marketplace - Plugins',
           })}
         />
-        <HeaderLayout
-          title={formatMessage({
-            id: 'global.marketplace',
-            defaultMessage: 'Marketplace',
-          })}
-          subtitle={formatMessage({
-            id: 'admin.pages.MarketPlacePage.subtitle',
-            defaultMessage: 'Get more out of Strapi',
-          })}
-          primaryAction={
-            <LinkButton
-              startIcon={<Upload />}
-              variant="tertiary"
-              href="https://market.strapi.io/submit-plugin"
-              onClick={() => trackUsage('didSubmitPlugin')}
-            >
-              {formatMessage({
-                id: 'admin.pages.MarketPlacePage.submit.plugin.link',
-                defaultMessage: 'Submit your plugin',
-              })}
-            </LinkButton>
-          }
-        />
+        <MarketplacePageHeader />
         <ContentLayout>
           <Box width="25%" paddingBottom={4}>
             <Searchbar
